refactor: added full authentication and authorization, crud routes and others

+++ app/controllers/users_controller.ts
@@ -1,7 +1,8 @@
-import { loginValidator } from '#validators/user'
+import { createUserValidator, loginValidator } from '#validators/user'
 import { inject } from '@adonisjs/core'
 import type { HttpContext } from '@adonisjs/core/http'
 import { UserDAO } from '../../contracts/user_dao.js'
+import { UserFile } from '../../contracts/user_file.js'
 
 @inject()
 export default class UsersController {
@@ -9,6 +10,34 @@ export default class UsersController {
 
   async loginRender({ view }: HttpContext) {
     return view.render('pages/login')
+  }
+
+  async homeRender({ response, view, session }: HttpContext) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    const userId = session.get('user').uid
+    const user = await this.userDAO.getUser(userId)
+    return view.render('pages/home', {
+      name: user.fullName.toUpperCase(),
+      course: user.course.toUpperCase(),
+      rgm: user.rgm.padStart(15, '0'),
+      cardCode: user.cardCode,
+      date: user.validUntil.toFormat('MM/yyyy'),
+    })
+  }
+
+  async usersRender({ response, view, session }: HttpContext) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    if (!(await this.userDAO.isUserAdmin(session.get('user').uid))) return response.redirect('/home')
+    const users = await this.userDAO.getUsers()
+    return view.render('pages/users', { users })
+  }
+
+  @inject()
+  async avatarFile({ response, request, session }: HttpContext, userFile: UserFile) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    const userId = request.param('id')
+    const avatarInfo = await this.userDAO.getUser(userId)
+    return response.download(await userFile.getFilePath(avatarInfo.avatarId, avatarInfo?.avatarExt ?? ''))
   }
 
   async login({ request, response, session }: HttpContext) {
@@ -24,21 +53,46 @@ export default class UsersController {
 
   async logout({ response, session }: HttpContext) {
     if (session.has('user')) {
-      const sessionData = session.get('user')
-      await this.userDAO.deleteUserToken(sessionData.uid, sessionData.tid)
+      if (await this.userDAO.verifyUserToken(session.get('user').token)) {
+        const sessionData = session.get('user')
+        await this.userDAO.deleteUserToken(sessionData.uid, sessionData.tid)
+      }
       session.forget('user')
     }
     return response.redirect('/login')
   }
 
-  async homeRender({ view, session }: HttpContext) {
-    const userId = session.get('user')?.uid ?? ''
-    const user = await this.userDAO.getUser(userId)
-    return view.render('pages/home', {
-      name: user.fullName.toUpperCase(),
-      course: user.course.toUpperCase(),
-      rgm: user.rgm.padStart(15, '0'),
-      date: user.validUntil.toFormat('MM/yyyy'),
-    })
+  async refresh({ request, response, session }: HttpContext) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    if (!(await this.userDAO.isUserAdmin(session.get('user').uid))) return response.redirect('/home')
+    const userId = request.param('id')
+    await this.userDAO.refreshUser(userId)
+    return response.redirect('/users')
+  }
+
+  @inject()
+  async create({ request, response, session }: HttpContext, userFile: UserFile) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    if (!(await this.userDAO.isUserAdmin(session.get('user').uid))) return response.redirect('/home')
+    const input = await request.validateUsing(createUserValidator)
+    const avatarId = (
+      await this.userDAO.createUser({
+        avatarExt: input.avatar.extname!,
+        email: input.email,
+        password: input.password,
+        fullName: input.fullName,
+        course: input.course,
+      })
+    ).avatarId
+    await userFile.persistFile(input.avatar, avatarId)
+    return response.redirect('/users')
+  }
+
+  async delete({ request, response, session }: HttpContext) {
+    if (!session.has('user') || !(await this.userDAO.verifyUserToken(session.get('user').token))) return response.redirect('/logout')
+    if (!(await this.userDAO.isUserAdmin(session.get('user').uid))) return response.redirect('/home')
+    const userId = request.param('id')
+    if (userId !== session.get('user').uid) await this.userDAO.deleteUser(userId)
+    return response.redirect('/users')
   }
 }
